requirements.txt
pandas==1.0.0
Wand==0.5.8
spacy==2.2.3

ipython==7.13.0
Pillow==7.0.0

pytesseract==0.3.3
